Update SummarisedExperiment

Updates or replaces the colData or rowData of a SummarisedExperiment.

Usage

annotate_se(
  summarised_experiment,
  sample_metadata,
  multiqc_path = NULL,
  gtf = FALSE,
  qc = TRUE,
  mode = "replace"
)

Arguments

summarised_experiment

the SummarisedExperiment to update

sample_metadata

a data.frame to replace or amend colData

multiqc_path

the path to a multiqc report from which to extract the QC metrics or gtf

gtf

a logical indicating whether to extract the gtf file from the multiqc report and replace gene annotations for rowData

qc

a logical indicating whether to extract the QC metrics from the multiqc report and replace or amend gene annotations to rowData

mode

one of c("replace", "amend"), should the provided data replace or amend the existing elements? Default: "replace"

Value

a SummarisedExperiment

Details

Genes and associated metadata come from the gtf logged in the multiqc report.

Sample metadata is matched by the library_id ID from the sample_metadata table and the QC metrics from the pipeline run. If the multiqc path refers to a full QC run (qc = TRUE set in run_nf_core_rnaseq(), the QC metrics come from {out_path}/multiqc/star_salmon/multiqc_data/multiqc_general_stats.txt. If a Salmon-only run (qc = FALSE set in run_nf_core_rnaseq()), the QC metrics come from {out_path}/multiqc/multiqc_data/multiqc_salmon.txt.

Author

Denis O'Meally