Minimal metadata for mouse samples

Download a template for new samples

Column	Description

sample_id	Unique identifier with the format MHO_XXXX that identifies a tissue sample with respect to the mouse, tissue type and sample date. Used to match samples across assays.

library_id	Integrative Genomics Core CLN with the format COHP_XXXXX. Matches sequence reads, QC metrics, raw counts and metadata for each sequencing library. Not unique as some libraries are split across multiple sequencing runs.

fastq_1	Full path of fastq1, unique in the table

fastq_2	Full path of fastq2, unique in the table, or blank for single-end reads

hdf5	Full path of h5 file, Cell Rangers's HDF5 Feature-Barcode Matrix Format

strandedness	Sequencing library protocol (reverse, forward or unstranded)

assay	one of mRNA, miRNA, scRNA

mouse_id	A unique 3 or 4 digit identifier for each mouse; database held by Kuo lab for AML mice and Zhang lab for CML.

tissue	Tissue type: PBMC peripheral blood mononuclear cells; BM bone marrow aspirant; BM_CKIT ckit+ flow sorted BM; PBMC_CKIT ckit+ flow sorted PBMC

timepoint	Timepoint label; formatting is inconsistent across cohorts

treatment	Experimental treatment group

genotype	Mouse genotype

sex	Sex (M/F)

dob	Date of birth, YYYY-MM-DD

cohort	Logical groups of samples, generally corresponding to experimental cohorts

batch	An arbitrary string to indicate samples in the same sequencing run; unique within cohorts

dod	Date of death, YYYY-MM-DD. NA if the mouse survived the experiment.

sample_date	Date the sample was collected, YYYY-MM-DD

percent_ckit	Percentage of c-KIT+ cells, measured by flow cytometry (CD117)

sample_weeks	Timepoint in weeks (sample_date - min(sample_date)). Post treatment chemo samples begin at week 0, pretreatment samples < 0.

age_at_start	Age at start of the 1st sample, in weeks (min(sample_date) - dob)

dead	Dead or alive at time of sampling; 0 = alive, 1 = dead

ref_dim1	reference dimension 1 - typically PC1; can be UMAP or any other dimension reduction coordinate.

ref_dim2	reference dimension 2 - typically PC2

qc_pass_mapping	mRNA samples only. TRUE if STAR uniquely mapped reads >= mapping_threshold (5% by default), or FALSE if not.

Details

The update_metadata_mmu() function assembles the metadata for all RNAseq libraries from AML and CML mice, by consolidating data scraped from multiple sequencing run sheets, directly from sequencing folders, emails, and so forth. The code is complex and ugly and undoubtedly some errors will have made it through.

Raise an issue on GitHub to report erroneous or missing records.

Author

Denis O'Meally