Datasets published to the Haemdata pin board

Lists the name and version of each dataset published to the Haemdata pinboard in the current release. Pins can be retrieved with the get_pin() function.

Usage

data(published_pins)

Format

A data.frame with the following columns:

pin_name

Name of the dataset on the pin board, including the file extension

version

Version of the pin to which the dataset belongs

Examples

published_pins |> knitr::kable()
#> 
#> 
#> |pin_name                                                 |version                |
#> |:--------------------------------------------------------|:----------------------|
#> |mmu_10x_mir142ko_GENCODEm28_HLT.rds                      |20231002T225122Z-25dd2 |
#> |mmu_10x_mir142ko_GENCODEm28_HLT.h5ad                     |20231002T225122Z-25dd2 |
#> |mmu_10x_aml2022_GENCODEm28_HLT.rds                       |20231002T224857Z-a9d58 |
#> |mmu_10x_aml2022_GENCODEm28_HLT.h5ad                      |20231002T224857Z-a9d58 |
#> |mmu_10x_blastcrisis_GENCODEm28_HLT.rds                   |20231002T230117Z-02ae7 |
#> |mmu_10x_blastcrisis_GENCODEm28_HLT.h5ad                  |20231002T230117Z-02ae7 |
#> |metadata_mmu.csv                                         |20230907T012227Z-d3324 |
#> |metadata_hsa.csv                                         |20231002T222804Z-79e11 |
#> |mmu_mirna_all_mice_miRBase22_mirtop_counts.csv           |20230907T012005Z-6a6f3 |
#> |mmu_mirna_all_mice_miRBase22_bowtie_cpm.csv              |20230907T012024Z-a1f46 |
#> |mmu_mirna_all_mice_miRBase22_mirna_qc.csv                |20230907T012042Z-94109 |
#> |mmu_mrna_all_mice_GENCODEm28_HLT_qc_1tpm_in_5samples.csv |20230907T012318Z-bdd1c |
#> |mmu_mrna_all_mice_GENCODEm28_HLT_qc.rds                  |20231002T231618Z-44f15 |
#> |mmu_mrna_techrep_GENCODEm28_HLT_qc_1tpm_in_5samples.csv  |20230907T012145Z-f2f01 |
#> |mmu_mrna_techrep_GENCODEm28_HLT_qc.rds                   |20231002T231723Z-590d0 |
#> |hsa_mrna_flt3_GENCODEr40_qc_1tpm_in_5samples.csv         |20230907T011903Z-7834c |
#> |hsa_mrna_flt3_GENCODEr40_qc.rds                          |20231002T231545Z-55bed |
#> |hsa_mrna_mds_GENCODEr40_qc_1tpm_in_5samples.csv          |20230907T012501Z-d385f |
#> |hsa_mrna_mds_GENCODEr40_qc.rds                           |20231002T231742Z-e284c |
#> |hsa_mrna_kim_GENCODEr40_qc_1tpm_in_5samples.csv          |20230907T012107Z-88cf0 |
#> |hsa_mrna_kim_GENCODEr40_qc.rds                           |20231002T231655Z-36654 |
#> |hsa_mrna_non_aml_GENCODEr40_qc_1tpm_in_5samples.csv      |20231002T231603Z-2b9ee |
#> |hsa_mrna_non_aml_GENCODEr40_qc.rds                       |20231002T231620Z-f806e |