Run nf-core/rnaseq pipeline on Apollo

Makes an sbatch job script for the nf-core/rnaseq pipeline and submits it to the Apollo cluster. The pipeline version is specified by the rnaseq_release parameter in the _targets.R file. The default parameters are used for the pipeline, in addition to --pseudo_aligner salmon and --deseq2_vst.

Makes an sbatch job script for the nf-core/smrnaseq pipeline and submits it to the Apollo cluster. The pipeline version is specified by the smrnaseq_release parameter in the _targets.R file. The default parameters are used for the pipeline.

Makes an sbatch job script for cutadapt and submits it to the Apollo cluster, one job per library.

Usage

run_nf_core_rnaseq(run_folder, sample_sheet, ref_genome, qc = TRUE)

run_nfcore_smrnaseq(
  run_folder,
  sample_sheet,
  clip_r1_3p = NULL,
  species = "mmu"
)

run_cutadapt_smrna(sample_sheet)

Arguments

run_folder

subfolder in which to run nextflow, keeps the pipeline from overwriting previous runs and enables the -resume feature of Nextflow.

sample_sheet

nf-core sample sheet with the columns sample_id, library_id, fastq_1

ref_genome

nf-core reference genome, see Reference Genomes for supported references

qc

run the full QC pipeline with STAR_Salmon, or Salmon pseudoalignment only, skipping all QC steps (default: TRUE)

clip_r1_3p

number of bases to trim from reads.

species

mmu or hsa. If mmu, additionally sets the 3' adapter to TCTGGAATTCTCGGGTGCCAAGGAACTCC. For hsa, the 3' adapter is auto-discovered.

Value

a path to the run script, or the multiqc report if it exists.

a path to the run script, or the multiqc report if it exists.

a data.frame with the columns library_id, fastq_1, ready to be used as input to the nf-core pipeline

Details

Execution of the pipeline is independent of the package build script, allowing asynchronous development. If no run_folder exists, it will be created along with a sample_sheet.csv and run_{ref_genome}.sh, which is submitted via sbatch. A successful run is detected by the presence of a a multiqc report. An existing run script wont be overwritten by this function so it must be deleted manually.

Execution of the pipeline is independent of the package build script, allowing asynchronous development. If no run_folder exists, it will be created along with a sample_sheet.csv and run_{ref_genome}.sh, which is submitted via sbatch. A successful run is detected by the presence of a a multiqc report. An existing run script wont be overwritten by this function so it must be deleted manually.

Execution of the pipeline is independent of the package build script, allowing asynchronous development. If no run_folder exists, it will be created along with a sample_sheet.csv and run_{ref_genome}.sh, which is submitted via sbatch. A successful run is detected by the presence of a a multiqc report. An existing run script wont be overwritten by this function so it must be deleted manually.

Author

Denis O'Meally