Perform PCA on a SummarisedExperiment

Perform PCA on the assay of a SummarisedExperiment object and make a biplot, pairs plot and correlation of metadata columns with PCs.

Usage

pca_se(summarised_experiment, assay = "abundance", col_by = NULL)

Arguments

summarised_experiment

a SummarisedExperiment object

assay

the assay to perform PCA on, Default: "abundance"

col_by

column name of colData to use for colouring samples in PCA plots. If none is given, samples will be coloured by the "condition", "treatment" or "library_id" column if any exists, in that order. Default: NULL

Value

a list of ggplot2 plots

Details

Could be rewritten, but this is a quick and dirty way to get something working using vignette("PCAtools", package = "PCAtools").

See also

Author

Denis O'Meally

Examples

if (FALSE) { # \dontrun{
if (interactive()) {
    se_pca(AML.mRNA.2016_qc_se)
}
} # }